Text Editor
Application created with C# and Windows Forms App (.NET Framework). It can save the text in a new file or you can access a file that already exists and you can modify the text, it also has many functions that allow you to play with the words in the text.